requirements.txt
uvicorn==0.23.1
google-generativeai==0.3.2
langchain==0.1.13
fastapi==0.85.1
langchain-google-genai==0.0.9
python-dotenv==1.0.1
requests==2.31.0
docx2txt==0.8
langchain-community==0.0.29
chromadb==0.3.29
python-docx==1.1.2